Fix imports of renamed files
